requirements.txt
pandas
mlflow==2.2.2
notebook
numpy
sklearn
matplotlib
python-box==6.0.2
pyYAML 
tqdm
ensure==1.0.2
joblib
types-pyYAML
Flask
Flask-Cors
python-dotenv==1.0.0
-e .